// packages/src/infra/persistence/mysql/prisma/generate/models/ScheduledGame.ts
export type ScheduledGameModel = runtime.Types.Result.DefaultSelection<Prisma.$ScheduledGamePayload>
// ...
export type $ScheduledGamePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ScheduledGame"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    date: Date
    category: string
    homeTeam: string
    awayTeam: string
    ballPark: string
    status: $Enums.GameStatus
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["scheduledGame"]>
  composites: {}
}